Fixed oversize chamfer, removed unecessary lip around rubber washer
Screws may push against and deform the displays if you torque them down too hard. The issue is the chamfer on the hole is too large. If a screw is torqued down hard enough to deform the plastic, then the tip of screw starts pushing into the display. It's not an issue on the 25cm trays because the displays are further from the screws.