Updates to Calibration Readme:
Specifically regarding the importation of opencv, the readme currently states to import opencv-python, however this does not include the proper files in order to import aruco, which exists in the extended library opencv-contrib-python.

I've also added an additional flag ", allow_pickle=True"to the .npz file import in the dualStereoChessboardCalibration.py file, located on line 77. We wasted a lot of time trying to figure out why numpy couldn't parse the .npz file in our python environments and this solved that issue.

Special thanks to Alina Christenbury and SafariMonkey from the northstar discord server for working with me on this.